adds comparison logic to post route
created a function that runs independent array calculation loop to be
called inside the loop for the full data array. function takes two
arrays and returns the total difference between them. full data array
loop runs each scores array through the totalDiff function and compares
the difference with the current lowest difference and returns the index
with the lowest difference.